Split serde logic from SerializedDAG
This allows us to actually NOT import SDK when SerializedDAG is used
in Core processes.
