+++ airflow-core/src/airflow/api/common/mark_tasks.py
@@ -76,10 +76,13 @@ def set_state(
     if not tasks:
         return []
 
-    task_dags = {task[0].dag if isinstance(task, tuple) else task.dag for task in tasks}
+    task_dags = {
+        (dag.dag_id if dag else None): dag
+        for dag in (task[0].dag if isinstance(task, tuple) else task.dag for task in tasks)
+    }
     if len(task_dags) > 1:
         raise ValueError(f"Received tasks from multiple DAGs: {task_dags}")
-    dag = next(iter(task_dags))
+    dag = next(iter(task_dags.values()))
     if dag is None:
         raise ValueError("Received tasks with no DAG")
     if not run_id:
