+++ airflow-core/tests/unit/models/test_dag.py
@@ -72,8 +72,9 @@ from airflow.sdk.definitions.asset import Asset, AssetAlias, AssetAll, AssetAny
 from airflow.sdk.definitions.callback import AsyncCallback
 from airflow.sdk.definitions.deadline import DeadlineAlert, DeadlineReference
 from airflow.sdk.definitions.param import Param
+from airflow.serialization.definitions.dag import SerializedDAG
 from airflow.serialization.encoders import coerce_to_core_timetable
-from airflow.serialization.serialized_objects import LazyDeserializedDAG, SerializedDAG
+from airflow.serialization.serialized_objects import LazyDeserializedDAG
 from airflow.task.trigger_rule import TriggerRule
 from airflow.timetables.base import DagRunInfo, DataInterval, TimeRestriction, Timetable
 from airflow.timetables.simple import (
@@ -391,10 +392,8 @@ class TestDag:
     def test_bulk_write_to_db(self, testing_dag_bundle):
         clear_db_dags()
         dags = [
-            SerializedDAG.deserialize_dag(
-                SerializedDAG.serialize_dag(
-                    DAG(f"dag-bulk-sync-{i}", schedule=None, start_date=DEFAULT_DATE, tags=["test-dag"])
-                )
+            create_scheduler_dag(
+                DAG(f"dag-bulk-sync-{i}", schedule=None, start_date=DEFAULT_DATE, tags=["test-dag"])
             )
             for i in range(4)
         ]
@@ -478,10 +477,8 @@ class TestDag:
         """
         clear_db_dags()
         dags = [
-            SerializedDAG.deserialize_dag(
-                SerializedDAG.serialize_dag(
-                    DAG(f"dag-bulk-sync-{i}", schedule=None, start_date=DEFAULT_DATE, tags=["test-dag"])
-                )
+            create_scheduler_dag(
+                DAG(f"dag-bulk-sync-{i}", schedule=None, start_date=DEFAULT_DATE, tags=["test-dag"])
             )
             for i in range(1)
         ]
@@ -509,10 +506,8 @@ class TestDag:
         """
         clear_db_dags()
         dags = [
-            SerializedDAG.deserialize_dag(
-                SerializedDAG.serialize_dag(
-                    DAG(f"dag-bulk-sync-{i}", schedule=None, start_date=DEFAULT_DATE, tags=["test-dag"])
-                )
+            create_scheduler_dag(
+                DAG(f"dag-bulk-sync-{i}", schedule=None, start_date=DEFAULT_DATE, tags=["test-dag"])
             )
             for i in range(4)
         ]
@@ -544,14 +539,8 @@ class TestDag:
         mock_active_runs_of_dags = mock.MagicMock(side_effect=DagRun.active_runs_of_dags)
         with mock.patch.object(DagRun, "active_runs_of_dags", mock_active_runs_of_dags):
             dags_null_timetable = [
-                SerializedDAG.deserialize_dag(
-                    SerializedDAG.serialize_dag(DAG("dag-interval-None", schedule=None, start_date=TEST_DATE))
-                ),
-                SerializedDAG.deserialize_dag(
-                    SerializedDAG.serialize_dag(
-                        DAG("dag-interval-test", schedule=interval, start_date=TEST_DATE)
-                    )
-                ),
+                create_scheduler_dag(DAG("dag-interval-None", schedule=None, start_date=TEST_DATE)),
+                create_scheduler_dag(DAG("dag-interval-test", schedule=interval, start_date=TEST_DATE)),
             ]
             SerializedDAG.bulk_write_to_db("testing", None, dags_null_timetable)
             if interval:
@@ -690,7 +679,7 @@ class TestDag:
         )
 
         session = settings.Session()
-        SerializedDAG.deserialize_dag(SerializedDAG.serialize_dag(dag1)).clear()
+        create_scheduler_dag(dag1).clear()
         SerializedDAG.bulk_write_to_db("testing", None, [dag1, dag2], session=session)
         session.commit()
         stored_assets = {x.uri: x for x in session.query(AssetModel).all()}
@@ -2756,12 +2745,12 @@ def test_iter_dagrun_infos_between_error(caplog):
 
     assert caplog.record_tuples == [
         (
-            "airflow.serialization.serialized_objects",
+            "airflow.serialization.definitions.dag",
             logging.ERROR,
             f"Failed to fetch run info after data interval {DataInterval(start, end)} for DAG {dag.dag_id!r}",
         ),
     ]
-    assert caplog.entries[0].get("exc_info") is not None, "should contain exception context"
+    assert caplog.entries[0].get("exception"), "should contain exception context"
 
 
 @pytest.mark.parametrize(
@@ -2940,7 +2929,7 @@ def test_create_dagrun_disallow_manual_to_use_automated_run_id(run_id_type: DagR
             f"A manual DAG run cannot use ID {run_id!r} since it is reserved for {run_id_type.value} runs"
         ),
     ):
-        SerializedDAG.deserialize_dag(SerializedDAG.serialize_dag(dag)).create_dagrun(
+        create_scheduler_dag(dag).create_dagrun(
             run_type=DagRunType.MANUAL,
             run_id=run_id,
             logical_date=DEFAULT_DATE,
