+++ airflow-core/tests/unit/serialization/test_dag_serialization.py
@@ -69,14 +69,15 @@ from airflow.sdk.definitions.operator_resources import Resources
 from airflow.sdk.definitions.param import Param, ParamsDict
 from airflow.security import permissions
 from airflow.serialization.definitions.assets import SerializedAssetUniqueKey
+from airflow.serialization.definitions.dag import SerializedDAG
 from airflow.serialization.definitions.notset import NOTSET
 from airflow.serialization.encoders import ensure_serialized_asset
 from airflow.serialization.enums import Encoding
 from airflow.serialization.json_schema import load_dag_schema_dict
 from airflow.serialization.serialized_objects import (
     BaseSerialization,
+    DagSerialization,
     SerializedBaseOperator,
-    SerializedDAG,
     SerializedParam,
     XComOperatorLink,
 )
@@ -459,7 +460,7 @@ def serialize_subprocess(queue, dag_folder):
     """Validate pickle in a subprocess."""
     dags, _ = collect_dags(dag_folder)
     for dag in dags.values():
-        queue.put(SerializedDAG.to_json(dag))
+        queue.put(DagSerialization.to_json(dag))
     queue.put(None)
 
 
@@ -512,8 +513,8 @@ class TestStringifiedDAGs:
             dags, import_errors = collect_dags()
         serialized_dags = {}
         for v in dags.values():
-            dag = SerializedDAG.to_dict(v)
-            SerializedDAG.validate_schema(dag)
+            dag = DagSerialization.to_dict(v)
+            DagSerialization.validate_schema(dag)
             serialized_dags[v.dag_id] = dag
 
         # Ignore some errors.
@@ -563,8 +564,8 @@ class TestStringifiedDAGs:
     def test_dag_serialization_to_timetable(self, timetable, serialized_timetable):
         """Verify a timetable-backed DAG is serialized correctly."""
         dag = get_timetable_based_simple_dag(timetable)
-        serialized_dag = SerializedDAG.to_dict(dag)
-        SerializedDAG.validate_schema(serialized_dag)
+        serialized_dag = DagSerialization.to_dict(dag)
+        DagSerialization.validate_schema(serialized_dag)
 
         expected = copy.deepcopy(serialized_simple_dag_ground_truth)
         expected["dag"]["timetable"] = serialized_timetable
@@ -584,8 +585,8 @@ class TestStringifiedDAGs:
         """Verify that an explicitly empty access_control dict is preserved."""
         dag = make_simple_dag()
         dag.access_control = {}
-        serialized_dag = SerializedDAG.to_dict(dag)
-        SerializedDAG.validate_schema(serialized_dag)
+        serialized_dag = DagSerialization.to_dict(dag)
+        DagSerialization.validate_schema(serialized_dag)
 
         assert serialized_dag["dag"]["access_control"] == {
             "__type": "dict",
@@ -597,7 +598,7 @@ class TestStringifiedDAGs:
         """Verify serialization fails without timetable registration."""
         dag = get_timetable_based_simple_dag(CustomSerializationTimetable("bar"))
         with pytest.raises(SerializationError) as ctx:
-            SerializedDAG.to_dict(dag)
+            DagSerialization.to_dict(dag)
 
         message = (
             "Failed to serialize DAG 'simple_dag': Timetable class "
@@ -655,7 +656,7 @@ class TestStringifiedDAGs:
             v = queue.get()
             if v is None:
                 break
-            dag = SerializedDAG.from_json(v)
+            dag = DagSerialization.from_json(v)
             assert isinstance(dag, SerializedDAG)
             stringified_dags[dag.dag_id] = dag
 
@@ -678,7 +679,7 @@ class TestStringifiedDAGs:
 
         # Verify deserialized DAGs.
         for dag in dags.values():
-            serialized_dag = SerializedDAG.from_json(SerializedDAG.to_json(dag))
+            serialized_dag = DagSerialization.from_json(DagSerialization.to_json(dag))
             self.validate_deserialized_dag(serialized_dag, dag)
 
         # Let's not be exact about this, but if everything fails to parse we should fail this test too
@@ -693,7 +694,7 @@ class TestStringifiedDAGs:
     def test_dag_roundtrip_from_timetable(self, timetable):
         """Verify a timetable-backed serialization can be deserialized."""
         dag = get_timetable_based_simple_dag(timetable)
-        roundtripped = SerializedDAG.from_json(SerializedDAG.to_json(dag))
+        roundtripped = DagSerialization.from_json(DagSerialization.to_json(dag))
         self.validate_deserialized_dag(roundtripped, dag)
 
     def validate_deserialized_dag(self, serialized_dag: SerializedDAG, dag: DAG):
@@ -902,7 +903,7 @@ class TestStringifiedDAGs:
         dag = DAG(dag_id="simple_dag", schedule=None, start_date=dag_start_date)
         BaseOperator(task_id="simple_task", dag=dag, start_date=task_start_date)
 
-        serialized_dag = SerializedDAG.to_dict(dag)
+        serialized_dag = DagSerialization.to_dict(dag)
         if not task_start_date or dag_start_date >= task_start_date:
             # If dag.start_date > task.start_date -> task.start_date=dag.start_date
             # because of the logic in dag.add_task()
@@ -910,7 +911,7 @@ class TestStringifiedDAGs:
         else:
             assert "start_date" in serialized_dag["dag"]["tasks"][0]["__var"]
 
-        dag = SerializedDAG.from_dict(serialized_dag)
+        dag = DagSerialization.from_dict(serialized_dag)
         simple_task = dag.task_dict["simple_task"]
         assert simple_task.start_date == expected_task_start_date
 
@@ -922,7 +923,7 @@ class TestStringifiedDAGs:
         ) as dag:
             BaseOperator(task_id="simple_task")
             # should not raise RuntimeError: dictionary changed size during iteration
-            SerializedDAG.to_dict(dag)
+            DagSerialization.to_dict(dag)
 
     @pytest.mark.parametrize(
         ("dag_end_date", "task_end_date", "expected_task_end_date"),
@@ -953,7 +954,7 @@ class TestStringifiedDAGs:
         )
         BaseOperator(task_id="simple_task", dag=dag, end_date=task_end_date)
 
-        serialized_dag = SerializedDAG.to_dict(dag)
+        serialized_dag = DagSerialization.to_dict(dag)
         if not task_end_date or dag_end_date <= task_end_date:
             # If dag.end_date < task.end_date -> task.end_date=dag.end_date
             # because of the logic in dag.add_task()
@@ -961,7 +962,7 @@ class TestStringifiedDAGs:
         else:
             assert "end_date" in serialized_dag["dag"]["tasks"][0]["__var"]
 
-        dag = SerializedDAG.from_dict(serialized_dag)
+        dag = DagSerialization.from_dict(serialized_dag)
         simple_task = dag.task_dict["simple_task"]
         assert simple_task.end_date == expected_task_end_date
 
@@ -1010,8 +1011,8 @@ class TestStringifiedDAGs:
                 "timetable": serialized_timetable,
             },
         }
-        SerializedDAG.validate_schema(serialized)
-        dag = SerializedDAG.from_dict(serialized)
+        DagSerialization.validate_schema(serialized)
+        dag = DagSerialization.from_dict(serialized)
         assert dag.timetable == expected_timetable
 
     @pytest.mark.parametrize(
@@ -1059,8 +1060,8 @@ class TestStringifiedDAGs:
                 "timetable": serialized_timetable,
             },
         }
-        SerializedDAG.validate_schema(serialized)
-        dag = SerializedDAG.from_dict(serialized)
+        DagSerialization.validate_schema(serialized)
+        dag = DagSerialization.from_dict(serialized)
         assert dag.timetable.summary == expected_timetable_summary
 
     def test_deserialization_timetable_unregistered(self):
@@ -1075,7 +1076,7 @@ class TestStringifiedDAGs:
                 "timetable": CUSTOM_TIMETABLE_SERIALIZED,
             },
         }
-        SerializedDAG.validate_schema(serialized)
+        DagSerialization.validate_schema(serialized)
         message = (
             "Timetable class "
             "'tests_common.test_utils.timetables.CustomSerializationTimetable' "
@@ -1084,7 +1085,7 @@ class TestStringifiedDAGs:
             "Please check the airflow best practices documentation."
         )
         with pytest.raises(ValueError, match=message):
-            SerializedDAG.from_dict(serialized)
+            DagSerialization.from_dict(serialized)
 
     @pytest.mark.parametrize(
         ("val", "expected"),
@@ -1110,10 +1111,10 @@ class TestStringifiedDAGs:
         ],
     )
     def test_roundtrip_relativedelta(self, val, expected):
-        serialized = SerializedDAG.serialize(val)
+        serialized = DagSerialization.serialize(val)
         assert serialized == expected
 
-        round_tripped = SerializedDAG.deserialize(serialized)
+        round_tripped = DagSerialization.deserialize(serialized)
         assert val == round_tripped
 
     @pytest.mark.parametrize(
@@ -1136,13 +1137,13 @@ class TestStringifiedDAGs:
         dag = DAG(dag_id="simple_dag", schedule=None, params=val)
         BaseOperator(task_id="simple_task", dag=dag, start_date=datetime(2019, 8, 1))
 
-        serialized_dag_json = SerializedDAG.to_json(dag)
+        serialized_dag_json = DagSerialization.to_json(dag)
 
         serialized_dag = json.loads(serialized_dag_json)
 
         assert "params" in serialized_dag["dag"]
 
-        deserialized_dag = SerializedDAG.from_dict(serialized_dag)
+        deserialized_dag = DagSerialization.from_dict(serialized_dag)
         deserialized_simple_task = deserialized_dag.task_dict["simple_task"]
         assert expected_val == deserialized_dag.params.dump()
         assert expected_val == deserialized_simple_task.params.dump()
@@ -1164,7 +1165,7 @@ class TestStringifiedDAGs:
         )
 
         with pytest.raises(SerializationError):
-            SerializedDAG.to_dict(dag)
+            DagSerialization.to_dict(dag)
 
         dag = DAG(dag_id="simple_dag", schedule=None)
         BaseOperator(
@@ -1189,10 +1190,10 @@ class TestStringifiedDAGs:
         Test to make sure that only native Param objects are being passed as dag or task params
         """
         sdk_dag = DAG(dag_id="simple_dag", schedule=None, params={"my_param": param})
-        serialized_json = SerializedDAG.to_json(sdk_dag)
+        serialized_json = DagSerialization.to_json(sdk_dag)
         serialized = json.loads(serialized_json)
-        SerializedDAG.validate_schema(serialized)
-        dag = SerializedDAG.from_dict(serialized)
+        DagSerialization.validate_schema(serialized)
+        dag = DagSerialization.from_dict(serialized)
 
         assert dag.params.get_param("my_param").value == param.value
         observed_param = dag.params.get_param("my_param")
@@ -1235,8 +1236,8 @@ class TestStringifiedDAGs:
             params=val,
             start_date=datetime(2019, 8, 1),
         )
-        serialized_dag = SerializedDAG.to_dict(dag)
-        deserialized_dag = SerializedDAG.from_dict(serialized_dag)
+        serialized_dag = DagSerialization.to_dict(dag)
+        deserialized_dag = DagSerialization.from_dict(serialized_dag)
 
         if val:
             assert "params" in serialized_dag["dag"]["tasks"][0]["__var"]
@@ -1290,10 +1291,10 @@ class TestStringifiedDAGs:
         with dag_maker(dag_id="simple_dag", start_date=test_date) as dag:
             CustomOperator(task_id="simple_task", bash_command=bash_command)
 
-        serialized_dag = SerializedDAG.to_dict(dag)
+        serialized_dag = DagSerialization.to_dict(dag)
         assert "bash_command" in serialized_dag["dag"]["tasks"][0]["__var"]
 
-        dag = SerializedDAG.from_dict(serialized_dag)
+        dag = DagSerialization.from_dict(serialized_dag)
         simple_task = dag.task_dict["simple_task"]
         assert getattr(simple_task, "bash_command") == bash_command
 
@@ -1435,8 +1436,8 @@ class TestStringifiedDAGs:
         with dag:
             BashOperator(task_id="test", bash_command=templated_field)
 
-        serialized_dag = SerializedDAG.to_dict(dag)
-        deserialized_dag = SerializedDAG.from_dict(serialized_dag)
+        serialized_dag = DagSerialization.to_dict(dag)
+        deserialized_dag = DagSerialization.from_dict(serialized_dag)
         deserialized_test_task = deserialized_dag.task_dict["test"]
         assert expected_field == getattr(deserialized_test_task, "bash_command")
 
@@ -1587,7 +1588,10 @@ class TestStringifiedDAGs:
             "ui_fgcolor": "#000",
         }
 
-        SerializedDAG._json_schema.validate(blob, _schema=load_dag_schema_dict()["definitions"]["operator"])
+        DagSerialization._json_schema.validate(
+            blob,
+            _schema=load_dag_schema_dict()["definitions"]["operator"],
+        )
         serialized_op = SerializedBaseOperator.deserialize_operator(blob)
         assert serialized_op.downstream_task_ids == {"foo"}
 
@@ -1602,9 +1606,9 @@ class TestStringifiedDAGs:
         with DAG("test_task_resources", schedule=None, start_date=logical_date) as dag:
             task = EmptyOperator(task_id=task_id, resources={"cpus": 0.1, "ram": 2048})
 
-        SerializedDAG.validate_schema(SerializedDAG.to_dict(dag))
+        DagSerialization.validate_schema(DagSerialization.to_dict(dag))
 
-        json_dag = SerializedDAG.from_json(SerializedDAG.to_json(dag))
+        json_dag = DagSerialization.from_json(DagSerialization.to_json(dag))
         deserialized_task = json_dag.get_task(task_id)
         assert deserialized_task.resources == task.resources
         assert isinstance(deserialized_task.resources, Resources)
@@ -1629,12 +1633,12 @@ class TestStringifiedDAGs:
             task1 >> group234
             group34 >> task5
 
-        dag_dict = SerializedDAG.to_dict(dag)
-        SerializedDAG.validate_schema(dag_dict)
-        json_dag = SerializedDAG.from_json(SerializedDAG.to_json(dag))
+        dag_dict = DagSerialization.to_dict(dag)
+        DagSerialization.validate_schema(dag_dict)
+        json_dag = DagSerialization.from_json(DagSerialization.to_json(dag))
         self.validate_deserialized_dag(json_dag, dag)
 
-        serialized_dag = SerializedDAG.deserialize_dag(SerializedDAG.serialize_dag(dag))
+        serialized_dag = DagSerialization.deserialize_dag(DagSerialization.serialize_dag(dag))
 
         assert serialized_dag.task_group.children
         assert serialized_dag.task_group.children.keys() == dag.task_group.children.keys()
@@ -1691,12 +1695,12 @@ class TestStringifiedDAGs:
                     EmptyOperator(task_id="task2")
                     EmptyOperator(task_id="teardown2").as_teardown()
 
-        dag_dict = SerializedDAG.to_dict(dag)
-        SerializedDAG.validate_schema(dag_dict)
-        json_dag = SerializedDAG.from_json(SerializedDAG.to_json(dag))
+        dag_dict = DagSerialization.to_dict(dag)
+        DagSerialization.validate_schema(dag_dict)
+        json_dag = DagSerialization.from_json(DagSerialization.to_json(dag))
         self.validate_deserialized_dag(json_dag, dag)
 
-        serialized_dag = SerializedDAG.deserialize_dag(SerializedDAG.serialize_dag(dag))
+        serialized_dag = DagSerialization.deserialize_dag(DagSerialization.serialize_dag(dag))
 
         self.assert_taskgroup_children(
             serialized_dag.task_group, dag.task_group, {"setup", "teardown", "group1"}
@@ -1738,12 +1742,12 @@ class TestStringifiedDAGs:
 
             mytask()
 
-        dag_dict = SerializedDAG.to_dict(dag)
-        SerializedDAG.validate_schema(dag_dict)
-        json_dag = SerializedDAG.from_json(SerializedDAG.to_json(dag))
+        dag_dict = DagSerialization.to_dict(dag)
+        DagSerialization.validate_schema(dag_dict)
+        json_dag = DagSerialization.from_json(DagSerialization.to_json(dag))
         self.validate_deserialized_dag(json_dag, dag)
 
-        serialized_dag = SerializedDAG.deserialize_dag(SerializedDAG.serialize_dag(dag))
+        serialized_dag = DagSerialization.deserialize_dag(DagSerialization.serialize_dag(dag))
         task = serialized_dag.task_group.children["mytask"]
         assert task.is_teardown is True
         assert task.on_failure_fail_dagrun is True
@@ -1755,10 +1759,10 @@ class TestStringifiedDAGs:
         )
         assert not dagbag.import_errors
         dag = dagbag.dags["example_dynamic_task_mapping"]
-        ser_dag = SerializedDAG.to_dict(dag)
+        ser_dag = DagSerialization.to_dict(dag)
         # We should not include `_is_sensor` most of the time (as it would be wasteful). Check we don't
         assert "_is_sensor" not in ser_dag["dag"]["tasks"][0]["__var"]
-        SerializedDAG.validate_schema(ser_dag)
+        DagSerialization.validate_schema(ser_dag)
 
     @pytest.mark.db_test
     def test_teardown_mapped_serialization(self, dag_maker):
@@ -1774,12 +1778,12 @@ class TestStringifiedDAGs:
         assert task.partial_kwargs["is_teardown"] is True
         assert task.partial_kwargs["on_failure_fail_dagrun"] is True
 
-        dag_dict = SerializedDAG.to_dict(dag)
-        SerializedDAG.validate_schema(dag_dict)
-        json_dag = SerializedDAG.from_json(SerializedDAG.to_json(dag))
+        dag_dict = DagSerialization.to_dict(dag)
+        DagSerialization.validate_schema(dag_dict)
+        json_dag = DagSerialization.from_json(DagSerialization.to_json(dag))
         self.validate_deserialized_dag(json_dag, dag)
 
-        serialized_dag = SerializedDAG.deserialize_dag(SerializedDAG.serialize_dag(dag))
+        serialized_dag = DagSerialization.deserialize_dag(DagSerialization.serialize_dag(dag))
         task = serialized_dag.task_group.children["mytask"]
         assert task.partial_kwargs["is_teardown"] is True
         assert task.partial_kwargs["on_failure_fail_dagrun"] is True
@@ -1831,7 +1835,7 @@ class TestStringifiedDAGs:
                 task2 = EmptyOperator(task_id="task2")
                 task1 >> task2
 
-            dag = SerializedDAG.to_dict(dag)
+            dag = DagSerialization.to_dict(dag)
             assert dag["dag"]["dag_dependencies"] == [
                 {
                     "source": "external_dag_id",
@@ -1873,7 +1877,7 @@ class TestStringifiedDAGs:
             for asset in testing_assets
         ]
 
-        dag = SerializedDAG.to_dict(dag)
+        dag = DagSerialization.to_dict(dag)
         actual = sorted(dag["dag"]["dag_dependencies"], key=lambda x: tuple(x.values()))
         expected = sorted(
             [
@@ -1975,7 +1979,7 @@ class TestStringifiedDAGs:
             for asset in testing_assets
         ]
 
-        dag = SerializedDAG.to_dict(dag)
+        dag = DagSerialization.to_dict(dag)
         actual = sorted(dag["dag"]["dag_dependencies"], key=lambda x: tuple(x.values()))
         expected = sorted(
             [
@@ -2052,7 +2056,7 @@ class TestStringifiedDAGs:
                     )
                 task1 >> task2
 
-            dag = SerializedDAG.to_dict(dag)
+            dag = DagSerialization.to_dict(dag)
             assert dag["dag"]["dag_dependencies"] == [
                 {
                     "source": "test_derived_dag_deps_trigger",
@@ -2153,12 +2157,12 @@ class TestStringifiedDAGs:
             task2 = EmptyOperator(task_id="task2")
             task1 >> Label("test label") >> task2
 
-        dag_dict = SerializedDAG.to_dict(dag)
-        SerializedDAG.validate_schema(dag_dict)
-        json_dag = SerializedDAG.from_json(SerializedDAG.to_json(dag))
+        dag_dict = DagSerialization.to_dict(dag)
+        DagSerialization.validate_schema(dag_dict)
+        json_dag = DagSerialization.from_json(DagSerialization.to_json(dag))
         self.validate_deserialized_dag(json_dag, dag)
 
-        serialized_dag = SerializedDAG.deserialize_dag(SerializedDAG.serialize_dag(dag))
+        serialized_dag = DagSerialization.deserialize_dag(DagSerialization.serialize_dag(dag))
 
         assert serialized_dag.edge_info == dag.edge_info
 
@@ -2220,13 +2224,13 @@ class TestStringifiedDAGs:
         )
         BaseOperator(task_id="simple_task", dag=dag, start_date=datetime(2019, 8, 1))
 
-        serialized_dag = SerializedDAG.to_dict(dag)
+        serialized_dag = DagSerialization.to_dict(dag)
         if expected_value:
             assert "has_on_success_callback" in serialized_dag["dag"]
         else:
             assert "has_on_success_callback" not in serialized_dag["dag"]
 
-        deserialized_dag = SerializedDAG.from_dict(serialized_dag)
+        deserialized_dag = DagSerialization.from_dict(serialized_dag)
 
         assert deserialized_dag.has_on_success_callback is expected_value
 
@@ -2252,13 +2256,13 @@ class TestStringifiedDAGs:
         )
         BaseOperator(task_id="simple_task", dag=dag, start_date=datetime(2019, 8, 1))
 
-        serialized_dag = SerializedDAG.to_dict(dag)
+        serialized_dag = DagSerialization.to_dict(dag)
         if expected_value:
             assert "has_on_failure_callback" in serialized_dag["dag"]
         else:
             assert "has_on_failure_callback" not in serialized_dag["dag"]
 
-        deserialized_dag = SerializedDAG.from_dict(serialized_dag)
+        deserialized_dag = DagSerialization.from_dict(serialized_dag)
 
         assert deserialized_dag.has_on_failure_callback is expected_value
 
@@ -2291,8 +2295,8 @@ class TestStringifiedDAGs:
                 **kwargs,
             )
             BaseOperator(task_id="simple_task", dag=dag, start_date=datetime(2019, 8, 1))
-            serialized_dag = SerializedDAG.to_dict(dag)
-            deserialized_dag = SerializedDAG.from_dict(serialized_dag)
+            serialized_dag = DagSerialization.to_dict(dag)
+            deserialized_dag = DagSerialization.from_dict(serialized_dag)
             assert deserialized_dag.disable_bundle_versioning is expected
 
     @pytest.mark.parametrize(
@@ -2341,7 +2345,7 @@ class TestStringifiedDAGs:
     )
     def test_serialized_objects_are_sorted(self, object_to_serialized, expected_output):
         """Test Serialized Sets are sorted while list and tuple preserve order"""
-        serialized_obj = SerializedDAG.serialize(object_to_serialized)
+        serialized_obj = DagSerialization.serialize(object_to_serialized)
         if isinstance(serialized_obj, dict) and "__type" in serialized_obj:
             serialized_obj = serialized_obj["__var"]
         assert serialized_obj == expected_output
@@ -2358,7 +2362,7 @@ class TestStringifiedDAGs:
                 "params": {"none": None, "str": "str", "dict": {"a": "b"}},
             },
         }
-        dag = SerializedDAG.from_dict(serialized)
+        dag = DagSerialization.from_dict(serialized)
 
         assert dag.params["none"] is None
         # After decoupling, server-side deserialization uses SerializedParam
@@ -2385,7 +2389,7 @@ class TestStringifiedDAGs:
                 },
             },
         }
-        dag = SerializedDAG.from_dict(serialized)
+        dag = DagSerialization.from_dict(serialized)
 
         param = dag.params.get_param("my_param")
         # After decoupling, server-side deserialization uses SerializedParam
@@ -2408,8 +2412,8 @@ class TestStringifiedDAGs:
                 "params": [["str", {"__class": "airflow.models.param.Param", "default": "str"}]],
             },
         }
-        SerializedDAG.validate_schema(serialized)
-        dag = SerializedDAG.from_dict(serialized)
+        DagSerialization.validate_schema(serialized)
+        dag = DagSerialization.from_dict(serialized)
 
         # After decoupling, server-side deserialization uses SerializedParam
         assert isinstance(dag.params.get_param("str"), SerializedParam)
@@ -2436,8 +2440,8 @@ class TestStringifiedDAGs:
                 ],
             },
         }
-        SerializedDAG.validate_schema(serialized)
-        dag = SerializedDAG.from_dict(serialized)
+        DagSerialization.validate_schema(serialized)
+        dag = DagSerialization.from_dict(serialized)
 
         assert dag.params["my_param"] == "a string value"
         param = dag.params.get_param("my_param")
@@ -2481,7 +2485,7 @@ class TestStringifiedDAGs:
                 )
             ),
         ):
-            SerializedDAG.to_dict(dag)
+            DagSerialization.to_dict(dag)
 
     @pytest.mark.db_test
     def test_start_trigger_args_in_serialized_dag(self):
@@ -2529,7 +2533,7 @@ class TestStringifiedDAGs:
             TestOperator(task_id="test_task_1")
             Test2Operator(task_id="test_task_2")
 
-        serialized_obj = SerializedDAG.to_dict(dag)
+        serialized_obj = DagSerialization.to_dict(dag)
         tasks = serialized_obj["dag"]["tasks"]
 
         assert tasks[0]["__var"]["start_trigger_args"] == {
@@ -2590,7 +2594,7 @@ def test_kubernetes_optional():
             module.BaseSerialization.from_dict(pod_override)
 
         # basic serialization should succeed
-        module.SerializedDAG.to_dict(make_simple_dag())
+        module.DagSerialization.to_dict(make_simple_dag())
 
 
 def test_operator_expand_serde():
@@ -2681,7 +2685,7 @@ def test_operator_expand_xcomarg_serde():
     xcom_ref = op.expand_input.value["arg2"]
     assert xcom_ref == _XComRef({"task_id": "op1", "key": XCOM_RETURN_KEY})
 
-    serialized_dag: DAG = SerializedDAG.from_dict(SerializedDAG.to_dict(dag))
+    serialized_dag: DAG = DagSerialization.from_dict(DagSerialization.to_dict(dag))
 
     xcom_arg = serialized_dag.task_dict["task_2"].expand_input.value["arg2"]
     assert isinstance(xcom_arg, SchedulerPlainXComArg)
@@ -2740,7 +2744,7 @@ def test_operator_expand_kwargs_literal_serde(strict):
         {"a": _XComRef({"task_id": "op1", "key": XCOM_RETURN_KEY})},
     ]
 
-    serialized_dag: DAG = SerializedDAG.from_dict(SerializedDAG.to_dict(dag))
+    serialized_dag: DAG = DagSerialization.from_dict(DagSerialization.to_dict(dag))
 
     resolved_expand_value = serialized_dag.task_dict["task_2"].expand_input.value
     assert resolved_expand_value == [
@@ -2787,7 +2791,7 @@ def test_operator_expand_kwargs_xcomarg_serde(strict):
     xcom_ref = op.expand_input.value
     assert xcom_ref == _XComRef({"task_id": "op1", "key": XCOM_RETURN_KEY})
 
-    serialized_dag: DAG = SerializedDAG.from_dict(SerializedDAG.to_dict(dag))
+    serialized_dag: DAG = DagSerialization.from_dict(DagSerialization.to_dict(dag))
 
     xcom_arg = serialized_dag.task_dict["task_2"].expand_input.value
     assert isinstance(xcom_arg, SchedulerPlainXComArg)
@@ -3059,7 +3063,7 @@ def test_mapped_task_group_serde():
         },
     )
 
-    serde_dag = SerializedDAG.deserialize_dag(ser_dag[Encoding.VAR])
+    serde_dag = DagSerialization.deserialize_dag(ser_dag[Encoding.VAR])
     serde_tg = serde_dag.task_group.children["tg"]
     assert isinstance(serde_tg, SerializedTaskGroup)
     assert serde_tg._expand_input == SchedulerDictOfListsExpandInput({"a": [".", ".."]})
@@ -3096,7 +3100,7 @@ def test_mapped_task_with_operator_extra_links_property():
         "_task_module": "unit.serialization.test_dag_serialization",
         "_is_mapped": True,
     }
-    deserialized_dag = SerializedDAG.deserialize_dag(serialized_dag[Encoding.VAR])
+    deserialized_dag = DagSerialization.deserialize_dag(serialized_dag[Encoding.VAR])
     # operator defined links have to be instances of XComOperatorLink
     assert deserialized_dag.task_dict["task"].operator_extra_links == [
         XComOperatorLink(name="airflow", xcom_key="_link_AirflowLink2")
@@ -3408,13 +3412,13 @@ def test_handle_v1_serdag():
         },
     ]
 
-    SerializedDAG.conversion_v1_to_v2(v1)
-    SerializedDAG.conversion_v2_to_v3(v1)
+    DagSerialization.conversion_v1_to_v2(v1)
+    DagSerialization.conversion_v2_to_v3(v1)
 
-    dag = SerializedDAG.from_dict(v1)
+    dag = DagSerialization.from_dict(v1)
 
     expected_sdag = copy.deepcopy(serialized_simple_dag_ground_truth)
-    expected = SerializedDAG.from_dict(expected_sdag)
+    expected = DagSerialization.from_dict(expected_sdag)
 
     fields_to_verify = set(vars(expected).keys()) - {
         "task_group",  # Tested separately
@@ -3616,10 +3620,10 @@ def test_handle_v2_serdag():
     }
 
     # Test that v2 DAGs can be deserialized without conversion
-    dag = SerializedDAG.from_dict(v2)
+    dag = DagSerialization.from_dict(v2)
 
     expected_sdag = copy.deepcopy(serialized_simple_dag_ground_truth)
-    expected = SerializedDAG.from_dict(expected_sdag)
+    expected = DagSerialization.from_dict(expected_sdag)
 
     fields_to_verify = set(vars(expected).keys()) - {
         "task_group",  # Tested separately
@@ -3664,11 +3668,11 @@ def test_dag_schema_defaults_optimization():
     )
 
     # Serialize and check exclusions
-    serialized = SerializedDAG.to_dict(dag_with_defaults)
+    serialized = DagSerialization.to_dict(dag_with_defaults)
     dag_data = serialized["dag"]
 
     # Schema default fields should be excluded
-    for field in SerializedDAG.get_schema_defaults("dag").keys():
+    for field in DagSerialization.get_schema_defaults("dag").keys():
         assert field not in dag_data, f"Schema default field '{field}' should be excluded"
 
     # None fields should also be excluded
@@ -3677,7 +3681,7 @@ def test_dag_schema_defaults_optimization():
         assert field not in dag_data, f"None field '{field}' should be excluded"
 
     # Test deserialization restores defaults correctly
-    deserialized_dag = SerializedDAG.from_dict(serialized)
+    deserialized_dag = DagSerialization.from_dict(serialized)
 
     # Verify schema defaults are restored
     assert deserialized_dag.catchup is False
@@ -3697,7 +3701,7 @@ def test_dag_schema_defaults_optimization():
         description="Test description",  # Non-None
     )
 
-    serialized_non_defaults = SerializedDAG.to_dict(dag_non_defaults)
+    serialized_non_defaults = DagSerialization.to_dict(dag_non_defaults)
     dag_non_defaults_data = serialized_non_defaults["dag"]
 
     # Non-default values should be included
@@ -3720,7 +3724,7 @@ def test_email_optimization_removes_email_attrs_when_email_empty():
             email_on_retry=True,  # This should be removed during serialization
         )
 
-    serialized_dag = SerializedDAG.to_dict(dag)
+    serialized_dag = DagSerialization.to_dict(dag)
     task_serialized = serialized_dag["dag"]["tasks"][0]["__var"]
     assert task_serialized is not None
 
@@ -3737,7 +3741,7 @@ def test_email_optimization_removes_email_attrs_when_email_empty():
             email_on_retry=True,
         )
 
-        serialized_dag_with_email = SerializedDAG.to_dict(dag_with_email)
+        serialized_dag_with_email = DagSerialization.to_dict(dag_with_email)
         task_with_email_serialized = serialized_dag_with_email["dag"]["tasks"][0]["__var"]
 
         assert task_with_email_serialized is not None
@@ -3984,10 +3988,10 @@ def test_weight_rule_absolute_serialization_deserialization():
 
         test_task()
 
-    serialized_dag = SerializedDAG.to_dict(dag)
+    serialized_dag = DagSerialization.to_dict(dag)
     assert serialized_dag["dag"]["tasks"][0]["__var"]["weight_rule"] == "absolute"
 
-    deserialized_dag = SerializedDAG.from_dict(serialized_dag)
+    deserialized_dag = DagSerialization.from_dict(serialized_dag)
 
     deserialized_task = deserialized_dag.task_dict["test_task"]
     assert isinstance(deserialized_task.weight_rule, _AbsolutePriorityWeightStrategy)
@@ -4132,7 +4136,7 @@ class TestDeserializationDefaultsResolution:
                 BashOperator(task_id="task1", bash_command="echo 1")
                 BashOperator(task_id="task2", bash_command="echo 2")
 
-            serialized = SerializedDAG.to_dict(dag)
+            serialized = DagSerialization.to_dict(dag)
 
             # Should have one client_defaults section for all tasks
             assert "client_defaults" in serialized
@@ -4142,7 +4146,7 @@ class TestDeserializationDefaultsResolution:
             client_defaults = serialized["client_defaults"]["tasks"]
 
             # Deserialize and check both tasks get the defaults
-            deserialized_dag = SerializedDAG.from_dict(serialized)
+            deserialized_dag = DagSerialization.from_dict(serialized)
             deserialized_task1 = deserialized_dag.get_task("task1")
             deserialized_task2 = deserialized_dag.get_task("task2")
 
@@ -4164,7 +4168,7 @@ class TestDeserializationDefaultsResolution:
                 BashOperator(task_id="task1", bash_command="echo 1")
                 BashOperator(task_id="task2", bash_command="echo 1", retries=2)
 
-            serialized = SerializedDAG.to_dict(dag)
+            serialized = DagSerialization.to_dict(dag)
 
             # verify client_defaults has retries=3
             assert "client_defaults" in serialized
@@ -4178,10 +4182,10 @@ class TestDeserializationDefaultsResolution:
             task2_data = serialized["dag"]["tasks"][1]["__var"]
             assert task2_data.get("retries", -1) == 2
 
-            deserialized_task1 = SerializedDAG.from_dict(serialized).get_task("task1")
+            deserialized_task1 = DagSerialization.from_dict(serialized).get_task("task1")
             assert deserialized_task1.retries == 0
 
-            deserialized_task2 = SerializedDAG.from_dict(serialized).get_task("task2")
+            deserialized_task2 = DagSerialization.from_dict(serialized).get_task("task2")
             assert deserialized_task2.retries == 2
 
 
@@ -4199,14 +4203,14 @@ class TestMappedOperatorSerializationAndClientDefaults:
                 ).expand(bash_command=["echo 1", "echo 2", "echo 3"])
 
             # Serialize the DAG
-            serialized_dag = SerializedDAG.to_dict(dag)
+            serialized_dag = DagSerialization.to_dict(dag)
 
             # Should have client_defaults section
             assert "client_defaults" in serialized_dag
             assert "tasks" in serialized_dag["client_defaults"]
 
             # Deserialize and check that client_defaults are applied
-            deserialized_dag = SerializedDAG.from_dict(serialized_dag)
+            deserialized_dag = DagSerialization.from_dict(serialized_dag)
             deserialized_task = deserialized_dag.get_task("mapped_task")
 
             # Verify it's still a MappedOperator
@@ -4265,7 +4269,7 @@ class TestMappedOperatorSerializationAndClientDefaults:
                     **task_config,
                 ).expand(bash_command=["echo 1", "echo 2", "echo 3"])
 
-            serialized_dag = SerializedDAG.to_dict(dag)
+            serialized_dag = DagSerialization.to_dict(dag)
             mapped_task_serialized = serialized_dag["dag"]["tasks"][0]["__var"]
 
             assert mapped_task_serialized is not None
@@ -4380,8 +4384,8 @@ class TestMappedOperatorSerializationAndClientDefaults:
             ).expand(bash_command=["echo 1", "echo 2"])
 
         # Serialize and deserialize the DAG
-        serialized_dag = SerializedDAG.to_dict(dag)
-        deserialized_dag = SerializedDAG.from_dict(serialized_dag)
+        serialized_dag = DagSerialization.to_dict(dag)
+        deserialized_dag = DagSerialization.from_dict(serialized_dag)
         deserialized_task = deserialized_dag.get_task("mapped_task")
 
         # Verify the task has correct values after round-trip
@@ -4441,7 +4445,7 @@ def test_dag_default_args_callbacks_serialization(callbacks, expected_has_flags,
     with DAG(dag_id="test_default_args_callbacks", default_args=default_args) as dag:
         BashOperator(task_id="task1", bash_command="echo 1", dag=dag)
 
-    serialized_dag_dict = SerializedDAG.serialize_dag(dag)
+    serialized_dag_dict = DagSerialization.serialize_dag(dag)
     default_args_dict = serialized_dag_dict["default_args"][Encoding.VAR]
 
     for flag in expected_has_flags:
@@ -4453,5 +4457,5 @@ def test_dag_default_args_callbacks_serialization(callbacks, expected_has_flags,
     assert default_args_dict["owner"] == "test_owner"
     assert default_args_dict["retries"] == 2
 
-    deserialized_dag = SerializedDAG.deserialize_dag(serialized_dag_dict)
+    deserialized_dag = DagSerialization.deserialize_dag(serialized_dag_dict)
     assert deserialized_dag.dag_id == "test_default_args_callbacks"
